+++ libs/wire-api-federation/src/Wire/API/Federation/Error.hs
@@ -44,6 +44,7 @@ federationErrorToWai (FederationCallFailure failure) = addErrorData $
         ("Unexpected method: " <> LT.fromStrict (T.decodeUtf8 mth))
     FederationClientStreamingUnsupported -> federationInvalidCall "Streaming unsupported"
     FederationClientOutwardError outwardErr -> federationRemoteError outwardErr
+    FederationClientInwardError inwardErr -> federationRemoteInwardError inwardErr
     FederationClientServantError (Servant.DecodeFailure msg _) -> federationInvalidBody msg
     FederationClientServantError (Servant.FailureResponse _ _) ->
       Wai.mkError unexpectedFederationResponseStatus "unknown-federation-error" "Unknown federation error"
@@ -130,6 +131,19 @@ federationUnavailable err =
     "federation-not-available"
     ("Local federator not available: " <> LT.fromStrict err)
 
+-- FUTUREWORK: Add Domain where the error occured to the WAI error responses here
+federationRemoteInwardError :: Proto.InwardError -> Wai.Error
+federationRemoteInwardError err = Wai.mkError status (LT.fromStrict label) (LT.fromStrict msg)
+  where
+    msg = Proto.inwardErrorMsg err
+    (status, label) = case Proto.inwardErrorType err of
+      Proto.IInvalidEndpoint -> (HTTP.Status 531 "Version Mismatch", "inward-invalid-endpoint")
+      Proto.IFederationDeniedByRemote -> (HTTP.Status 532 "Federation Denied", "federation-denied-by-remote")
+      Proto.IInvalidDomain -> (unexpectedFederationResponseStatus, "invalid-origin-domain")
+      Proto.IForbiddenEndpoint -> (unexpectedFederationResponseStatus, "forbidden-endpoint")
+      Proto.IOther -> (unexpectedFederationResponseStatus, "inward-other")
+
+-- FUTUREWORK: Add Domain where the error occured to the WAI error responses here
 federationRemoteError :: Proto.OutwardError -> Wai.Error
 federationRemoteError err = Wai.mkError status (LT.fromStrict label) (LT.fromStrict msg)
   where