+++ CHANGELOG.md
@@ -38,6 +38,13 @@
   added to a new conversation via `POST /federation/register-conversation` (#1622).
 * [Federation] Adjust scripts under ./hack/federation to work with recent changes to the federation API (#1632).
 * Refactored Proteus endpoint to work with qualified users (#1634).
+* Refactored Federator InternalServer (#1637)
+
+### Internal Federation API changes
+
+(federation should not be in use anywhere yet, so this should not have any impact)
+
+* Breaking change on InwardResponse and OutwardResponse in router.proto for improved error handling (#1637)
 
 ## Documentation